implemented use rhandling in queries for movements and categories

+++ src/main/java/h4rk/finance/repository/CatsRepository.java
@@ -19,27 +19,32 @@ public class CatsRepository {
     @Autowired
     private JdbcTemplate jdbcTemplate;
 
-    public List<Cat> getCats() {
-        return jdbcTemplate.query("SELECT * FROM cat", (rs, rowNum) -> new Cat(rs.getLong("cat_id"), rs.getString("name"), rs.getString("description"), rs.getShort("cat_type")));
+    public List<Cat> getCats(long userId) {
+        return jdbcTemplate.query("SELECT * FROM cat WHERE user_id = ?", (rs, rowNum) -> new Cat(rs.getLong("cat_id"), rs.getString("name"), rs.getString("description"), rs.getShort("cat_type")), userId);
     }
 
-    public Cat postCat(Cat cat) {
+    public Cat postCat(Cat cat, long userId) {
 		KeyHolder keyHolder = new GeneratedKeyHolder();
 
     	jdbcTemplate.update(connection -> {
         PreparedStatement ps = connection
-          .prepareStatement("INSERT INTO cat (name, description, cat_type) VALUES (?, ?, ?)", Statement.RETURN_GENERATED_KEYS);
+          .prepareStatement("INSERT INTO cat (name, description, cat_type, user_id) VALUES (?, ?, ?, ?)", Statement.RETURN_GENERATED_KEYS);
           ps.setString(1, cat.getName());
 		  ps.setString(2, cat.getDescription());
 		  ps.setShort(3, cat.getType());
+		  ps.setLong(4, userId);
           return ps;
         }, keyHolder);
 		cat.setId(keyHolder.getKey().longValue());
 		return cat;
 	}
 
-    public int deleteCat(long id) {
-        return jdbcTemplate.update("DELETE FROM cat WHERE cat_id = ?", id);
+    public int deleteCat(long id, long userId) {
+        return jdbcTemplate.update("DELETE FROM cat WHERE cat_id = ? AND user_id = ?", id, userId);
     }
+
+	public Cat getCatById(Long id, long userId) {
+		return jdbcTemplate.queryForObject("SELECT * FROM cat WHERE cat_id = ? AND user_id = ?", (rs, rowNum) -> new Cat(rs.getLong("cat_id"), rs.getString("name"), rs.getString("description"), rs.getShort("cat_type")), id, userId);
+	}
     
 }
